# openapi_current.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kreeda API",
    "description": "Kreeda - Digital Cricket Scorekeeping Platform",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/matches/matches": {
      "get": {
        "tags": [
          "cricket-integrity"
        ],
        "summary": "Get Matches",
        "description": "Get all matches for the user",
        "operationId": "get_matches_api_v1_matches_matches_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MatchResponseBasic"
                  },
                  "type": "array",
                  "title": "Response Get Matches Api V1 Matches Matches Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "cricket-integrity"
        ],
        "summary": "Create Match",
        "description": "Create a new cricket match",
        "operationId": "create_match_api_v1_matches_matches_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CricketMatchCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchResponseBasic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/matches/matches/{match_id}": {
      "get": {
        "tags": [
          "cricket-integrity"
        ],
        "summary": "Get Match",
        "description": "Get match by ID",
        "operationId": "get_match_api_v1_matches_matches__match_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "match_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Match Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchResponseBasic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/matches/matches/{match_id}/scorers": {
      "post": {
        "tags": [
          "cricket-integrity"
        ],
        "summary": "Assign Match Scorers",
        "description": "Assign official scorers for a match",
        "operationId": "assign_match_scorers_api_v1_matches_matches__match_id__scorers_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "match_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Match Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScorerAssignment"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/matches/matches/{match_id}/scoring-status": {
      "get": {
        "tags": [
          "cricket-integrity"
        ],
        "summary": "Get Match Scoring Status",
        "description": "Get scoring status and integrity information for a match",
        "operationId": "get_match_scoring_status_api_v1_matches_matches__match_id__scoring_status_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "match_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Match Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/matches/matches/{match_id}/balls": {
      "post": {
        "tags": [
          "cricket-integrity"
        ],
        "summary": "Record Ball With Integrity",
        "description": "Record a ball with integrity verification",
        "operationId": "record_ball_with_integrity_api_v1_matches_matches__match_id__balls_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "match_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Match Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BallRecord"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/matches/matches/{match_id}/disputes/{innings}/{over_number}/{ball_number}/resolve": {
      "post": {
        "tags": [
          "cricket-integrity"
        ],
        "summary": "Resolve Scoring Dispute",
        "description": "Resolve a scoring dispute (umpire/referee only)",
        "operationId": "resolve_scoring_dispute_api_v1_matches_matches__match_id__disputes__innings___over_number___ball_number__resolve_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "match_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Match Id"
            }
          },
          {
            "name": "innings",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Innings"
            }
          },
          {
            "name": "over_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Over Number"
            }
          },
          {
            "name": "ball_number",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Ball Number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "title": "Resolution Data"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/matches/matches/{match_id}/scorecard": {
      "get": {
        "tags": [
          "cricket-integrity"
        ],
        "summary": "Get Scorecard",
        "description": "Get match scorecard (only verified balls)",
        "operationId": "get_scorecard_api_v1_matches_matches__match_id__scorecard_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "match_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Match Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/matches/matches/{match_id}/balls/legacy": {
      "post": {
        "tags": [
          "cricket-integrity"
        ],
        "summary": "Record Ball Legacy",
        "description": "Legacy ball recording without integrity checks (for testing only)",
        "operationId": "record_ball_legacy_api_v1_matches_matches__match_id__balls_legacy_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "match_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Match Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BallRecord"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BallResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BallRecord": {
        "properties": {
          "over_number": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Over Number"
          },
          "ball_number": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Ball Number"
          },
          "bowler_id": {
            "type": "string",
            "format": "uuid",
            "title": "Bowler Id"
          },
          "batsman_striker_id": {
            "type": "string",
            "format": "uuid",
            "title": "Batsman Striker Id"
          },
          "batsman_non_striker_id": {
            "type": "string",
            "format": "uuid",
            "title": "Batsman Non Striker Id"
          },
          "runs_scored": {
            "type": "integer",
            "maximum": 6.0,
            "minimum": 0.0,
            "title": "Runs Scored",
            "default": 0
          },
          "extras": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Extras",
            "default": 0
          },
          "ball_type": {
            "type": "string",
            "pattern": "^(legal|wide|no_ball|bye|leg_bye)$",
            "title": "Ball Type",
            "default": "legal"
          },
          "is_wicket": {
            "type": "boolean",
            "title": "Is Wicket",
            "default": false
          },
          "wicket_type": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(bowled|caught|lbw|run_out|stumped)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wicket Type"
          },
          "dismissed_player_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dismissed Player Id"
          },
          "is_boundary": {
            "type": "boolean",
            "title": "Is Boundary",
            "default": false
          },
          "boundary_type": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(four|six)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Boundary Type"
          }
        },
        "type": "object",
        "required": [
          "over_number",
          "ball_number",
          "bowler_id",
          "batsman_striker_id",
          "batsman_non_striker_id"
        ],
        "title": "BallRecord"
      },
...
      "CricketMatchCreate": {
        "properties": {
          "team_a_id": {
            "type": "string",
            "format": "uuid",
            "title": "Team A Id"
          },
          "team_b_id": {
            "type": "string",
            "format": "uuid",
            "title": "Team B Id"
          },
          "overs_per_innings": {
            "type": "integer",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Overs Per Innings",
            "default": 20
          },
          "venue": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Venue"
          },
          "match_date": {
            "type": "string",
            "format": "date-time",
            "title": "Match Date"
          }
        },
        "type": "object",
        "required": [
          "team_a_id",
          "team_b_id",
          "venue",
          "match_date"
        ],
        "title": "CricketMatchCreate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MatchResponseBasic": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "team_a_id": {
            "type": "string",
            "format": "uuid",
            "title": "Team A Id"
          },
          "team_b_id": {
            "type": "string",
            "format": "uuid",
            "title": "Team B Id"
          },
          "overs_per_innings": {
            "type": "integer",
            "title": "Overs Per Innings"
          },
          "venue": {
            "type": "string",
            "title": "Venue"
          },
          "match_date": {
            "type": "string",
            "format": "date-time",
            "title": "Match Date"
          },
          "created_by_id": {
            "type": "string",
            "format": "uuid",
            "title": "Created By Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "current_innings": {
            "type": "integer",
            "title": "Current Innings"
          },
          "team_a_score": {
            "type": "integer",
            "title": "Team A Score"
          },
          "team_a_wickets": {
            "type": "integer",
            "title": "Team A Wickets"
          },
          "team_a_overs": {
            "type": "number",
            "title": "Team A Overs"
          },
          "team_b_score": {
            "type": "integer",
            "title": "Team B Score"
          },
          "team_b_wickets": {
            "type": "integer",
            "title": "Team B Wickets"
          },
          "team_b_overs": {
            "type": "number",
            "title": "Team B Overs"
          },
          "toss_winner_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Toss Winner Id"
          },
          "toss_decision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Toss Decision"
          },
          "winner_team_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Winner Team Id"
          },
          "result_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Description"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "team_a_id",
          "team_b_id",
          "overs_per_innings",
          "venue",
          "match_date",
          "created_by_id",
          "status",
          "current_innings",
          "team_a_score",
          "team_a_wickets",
          "team_a_overs",
          "team_b_score",
          "team_b_wickets",
          "team_b_overs",
          "created_at",
          "updated_at"
        ],
        "title": "MatchResponseBasic",
        "description": "Basic match response without relationship data to prevent async loading issues"
      },
      "ScorerAssignment": {
        "properties": {
          "team_a_scorer_id": {
            "type": "string",
            "format": "uuid",
            "title": "Team A Scorer Id",
            "description": "User ID for Team A scorer"
          },
          "team_b_scorer_id": {
            "type": "string",
            "format": "uuid",
            "title": "Team B Scorer Id",
            "description": "User ID for Team B scorer"
          },
          "umpire_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Umpire Id",
            "description": "Optional umpire/referee user ID"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "match_id": {
            "type": "string",
            "format": "uuid",
            "title": "Match Id"
          },
          "innings": {
            "type": "integer",
            "title": "Innings"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "team_a_scorer_id",
          "team_b_scorer_id",
          "id",
          "match_id",
          "innings",
          "created_at"
        ],
        "title": "ScorerAssignment",
        "description": "Schema for assigning scorers to a match"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}